### 2018-03-09 #609 @whedon
[PRE REVIEW]: Open OnDemand: A web-based client portal for HPC centers
**Submitting author:** @ericfranz (<a href="http://orcid.org/0000-0002-9662-412X">Eric Franz</a>)
**Repository:** <a href="https://github.com/OSC/Open-OnDemand" target ="_blank">https://github.com/OSC/Open-OnDemand</a>
**Version:** v1.2
