### 2018-03-09 #609 @arfon
@danielskatz - would you be willing to edit this submission for JOSS?
