### 2018-03-12 #609 @danielskatz
@smgallo - would you be willing to review this submission to JOSS?
