### 2018-03-14 #622 @whedon
[REVIEW]: Open OnDemand: A web-based client portal for HPC centers
**Submitting author:** @ericfranz (<a href="http://orcid.org/0000-0002-9662-412X">Eric Franz</a>)
**Repository:** <a href="https://github.com/OSC/Open-OnDemand" target ="_blank">https://github.com/OSC/Open-OnDemand</a>
**Version:** v1.2
**Reviewer1:** @marpierc
**Reviewer2:** @smgallo
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.6265514.v1" target="_blank">10.6084/m9.figshare.6265514.v1</a>
