### 2018-03-14 #609 @danielskatz
@smgallo - great - having 2 reviewers is much better than 1.  See https://github.com/openjournals/joss-reviews/issues/622 which I'll finish setting up in 5 or so more minutes.
