### 2018-03-14 #622 @danielskatz
My one initial comment is that the paper doesn't have any references - the reviewers might think about any things (papers, systems) that should be referenced...
