### 2018-03-14 OSC/Open-OnDemand#30 @ericfranz
JOSS: Generated PDF did not pull references from paper.bib
@danielskatz [you wrote](https://github.com/openjournals/joss-reviews/issues/622#issuecomment-373012555):

>My one initial comment is that the paper doesn't have any references - the reviewers might think about any things (papers, systems) that should be referenced...

Two references are in the root of this repo at [paper.bib](https://github.com/OSC/Open-OnDemand/blob/556c0068dc89c462a5a2555aac277465e742391a/paper.bib).

Following the [author guidelines](http://joss.theoj.org/about#author_guidelines) with the example paper.md and paper.bib, I left `# References` blank and added those via paper.bib. Two references were from papers in ACM, so I exported the BibTeX format and added them to paper.bib. We could definitely expand the list references if appropriate.
