### 2018-03-14 OSC/Open-OnDemand#30 @danielskatz
Items in paper.bib only get pulled into the paper if they are referenced in the body( of paper.md)

Is this the case here?
