### 2018-03-14 OSC/Open-OnDemand#30 @ericfranz
I understand. No the paper.md does not directly reference the references. I'll pass this info internally and see about updating the paper.md as appropriate.

Also, is this the appropriate method of communication? Using GitHub issues on this repo? Or should I wait till more of the review work is complete and reviewers start opening issues on this repo?
