### 2018-03-14 OSC/Open-OnDemand#30 @danielskatz
It's really up to the reviewers to bring up issues.  We suggest that they generally discuss stuff here, and in the review issue (https://github.com/openjournals/joss-reviews/issues/622) point to issues here, and mark when they are closed.  I'll do that now.
