### 2018-03-14 #622 @danielskatz
discussing the references in https://github.com/openjournals/joss-reviews/issues/622
