### 2018-03-26 bcb7411
add references in paper
Fixes #30
