+++ paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{Hudak:2013:OOW:2484762.2484780,
+@inproceedings{Hudak2013,
  author = {Hudak, David E. and Bitterman, Thomas and Carey, Patricia and Johnson, Douglas and Franz, Eric and Brady, Shaun and Diwan, Piyush},
  title = {OSC OnDemand: A Web Platform Integrating Access to HPC Systems, Web and VNC Applications},
  booktitle = {Proceedings of the Conference on Extreme Science and Engineering Discovery Environment: Gateway to Discovery},
@@ -17,7 +17,7 @@
  keywords = {OpenID, REST, cyberinfrastructure, high performance computing, virtual organizations, web platform},
 }
 
-@inproceedings{Hudak:2016:OOT:2949550.2949644,
+@inproceedings{Hudak2016,
  author = {Hudak, David E. and Johnson, Douglas and Nicklas, Jeremy and Franz, Eric and McMichael, Brian and Gohar, Basil},
  title = {Open OnDemand: Transforming Computational Science Through Omnidisciplinary Software Cyberinfrastructure},
  booktitle = {Proceedings of the XSEDE16 Conference on Diversity, Big Data, and Science at Scale},
