+++ paper.md
@@ -37,8 +37,31 @@ bibliography: paper.bib
 
 # Summary
 
-The web has become the dominant access mechanism for remote compute services in every computing area except high-performance computing (HPC). Accessing HPC resources, either at the campus or national level typically requires advanced knowledge of Linux, familiarity with command-line interfaces and installation and configuration of custom client software (e.g., Secure Shell (SSH) and Virtual Network Computing (VNC)). These additional requirements create an accessibility gap for HPC. To help address this gap we have created the Open OnDemand Project, an open-source software project based on the proven Ohio Supercomputer Center (OSC) OnDemand platform, to allow HPC centers to provide advanced web and graphical interfaces for their users.
+The web has become the dominant access mechanism for remote compute services in
+every computing area except high-performance computing (HPC). Accessing HPC
+resources, either at the campus or national level typically requires advanced
+knowledge of Linux, familiarity with command-line interfaces and installation
+and configuration of custom client software (e.g., Secure Shell (SSH) and
+Virtual Network Computing (VNC)). These additional requirements create an
+accessibility gap for HPC. To help address this gap we have created the Open
+OnDemand Project [@Hudak2016], an open-source software project based on the
+proven Ohio Supercomputer Center (OSC) OnDemand platform [@Hudak2013], to allow
+HPC centers to provide advanced web and graphical interfaces for their users.
 
-Open OnDemand is the result of substantial development and integration efforts in four key areas.  (1) The per-user NGINX (PUN) architecture including federated authentication using CILogon, Apache-based web proxy, per-user NGINX configuration, and Unix domain sockets for secure server-side communication between the proxy and each PUN.  The PUN architecture is an original contribution of the project. (2) The file browser and file editor which, though originally based on an existing open source project, have been extensively modified. (3) The terminal, created by integrating an existing open source project with minimal effort.  (4) Accessibility Apps (Dashboard, Job Constructor, Job Status, System Status, VDI and iHPC apps) built using the Rails-based AweSim AppKit (which was developed by this team on a previous project and leveraged here with minor modifications). The AweSim AppKit allows for the development of both workflow and interactive applications and includes mechanisms for user-based app creation, app sharing and app publishing.  The AppKit technology is included as part of the Open OnDemand project.
+Open OnDemand is the result of substantial development and integration efforts
+in four key areas. (1) The per-user NGINX (PUN) architecture including
+federated authentication using CILogon, Apache-based web proxy, per-user NGINX
+configuration, and Unix domain sockets for secure server-side communication
+between the proxy and each PUN. The PUN architecture is an original
+contribution of the project. (2) The file browser and file editor which, though
+originally based on an existing open source project, have been extensively
+modified. (3) The terminal, created by integrating an existing open source
+project with minimal effort. (4) Accessibility Apps (Dashboard, Job
+Constructor, Job Status, System Status, VDI and iHPC apps) built using the
+Rails-based AweSim AppKit (which was developed by this team on a previous
+project and leveraged here with minor modifications). The AweSim AppKit allows
+for the development of both workflow and interactive applications and includes
+mechanisms for user-based app creation, app sharing and app publishing. The
+AppKit technology is included as part of the Open OnDemand project.
 
 # References
