### 2018-03-26 OSC/Open-OnDemand#30 @ericfranz
@danielskatz if the generated PDF still doesn't have the references after this change, or if more references are needed, let us know and we can reopen this issue.
