### 2018-03-27 OSC/Open-OnDemand#30 @danielskatz
I regenerated the PDF in https://github.com/openjournals/joss-reviews/issues/622 and it now has the review - thanks
