### 2018-03-27 #622 @smgallo
@danielskatz

OOD is an excellent package and fills a need for HPC users by simplifying and improving access to complicated HPC resources.  HPC centers are able to provide their users with simplified access to resources while making it easier to submit and manage both jobs and data. Single sign-on is supported making it easier to integrate with center or campus-wide authentication systems.  Because OOD runs in a web-browser, users do not need to install additional software on their computers to access HPC resources.

Our center recently installed OOD to serve novice and light HPC users and has found that even experienced users have said that using OOD is "so much easier" than logging in via the command line to manage their compute jobs and data. In addition, faculty who use our HPC resources to teach courses are entheuiastic about OOD.  Over the years, we had implemented a subset of the functionality that OOD provides to support the needs of our users, but having all of this functionality in a single integrated framework is great.  There is an active mailing list which is available for general questions. Multiple authentication methods are supported via Apache modules including shibboleth, LDAP, and CiLogin. We have implemented CiLogin backed by FreeIPA for access to our resources.

OOD supports a number of useful applications including a job status viewer, file manager and editor, shell tool, and an interactive desktop. In addition, there are interactive desktop applications that are supported such as MATLAb, Jupyter, and R Studio, all of which are desired by end users. Development of custom applications is [documented](https://osc.github.io/ood-documentation/master/app-development.html) (with examples). The interactive desktop application has proven especially useful and allowed us to deprecate a licensed application that was previously provided to our users. A very useful feature of OOD are the Per User NGIX Processes [(PUN)](https://osc.github.io/ood-documentation/master/infrastructure/nginx-stage.html). This allows users to run their applications via the web server in a process that they own (rather than the web server owning the process) making system and data security far simpler.

## General Comments

1. An RPM install is not available but it would be great to have this (along with an upgrade path from a source install). Is an RPM (or other package manager) version planned?
1. Individual apps are contained in their own git repos. While this makes extra work when installing the apps, it is nicely compartmentalized.

## Versioning

The versioning is a bit confusing and could use clarification in the documentation. OOD is made up of several components including infrastructure and applications. The infrastructure component versions do not appear to be to match the 1.2 version of OOD as a whole and the only application that I could find with a recent 1.23 version number was the dashboard. Other applications seem to have their own version tags but are now far ahead of version 1.2. For example, the `myjobs` application is now at version 2.8.2 with 1.2 released almost 2 years ago.

Thanks to CCR staff members @aebruno, @dsajdak, and Cynthia Cornelius at [CCR](http://www.buffalo.edu/ccr.html) for their input. CCR has installed Open on Demand and made it available to our users.
