### 2018-03-27 #622 @danielskatz
Thanks @smgallo - it looks like there are some issues you raised for @ericfranz to respond to or resolve.  Additionally, I see you did not check the automated tests item, or discuss testing in your comment above.  Can you explain?
