### 2018-04-02 #622 @danielskatz
@ericfranz - Do you have a response to @smgallo's comments?  You can certainly address them before @marpierc's review
