### 2018-04-02 OSC/Open-OnDemand#31 @ericfranz
Add OnDemand Roadmap to website
Related to https://github.com/OSC/Open-OnDemand/issues/27

We don't advertise well what we are working on. If we had a well advertised project roadmap it would be easier for partners and others interested to see what we are planning.

Example:

We have been working internally on an OnDemand rpm for easier installation. Currently it documented in the develop branch of the documentation at https://osc.github.io/ood-documentation/develop/installation/install-software.html and is available in https://yum.osc.edu/ondemand/latest/web/el7/x86_64/. When 1.3 is released, we will have rpms under https://yum.osc.edu/ondemand/stable for both el6 and el7. Besides mentioning our intention to do this in one of the final webinar slides, no one knows what we are currently working on.

The Helix project has an interesting looking roadmap page that we could use for inspiration: https://usehelix.com/roadmap. There may be others.
