### 2018-04-02 OSC/Open-OnDemand#32 @ericfranz
Clarify versioning of the packages that go together with OnDemand
Related to https://github.com/OSC/Open-OnDemand/issues/27

Currently OnDemand is made up of many individual packages in GitHub, each with their own version. We advertise the various versions of the repos of OnDemand version 1.0, 1.1, 1.2, etc. in the Release Notes section of the documentation. Also, the version of the main OnDemand rpm and the version of the ood-apps-installer corresponds to the OnDemand version (i.e. 1.1, 1.2, 1.3) and those act as a manifest for the versions of the other components installed.

This appears to be difficult to find, however. We may want to find a more prominent location for these two tables:

https://osc.github.io/ood-documentation/release-1.2/release-notes/v1.2-release-notes.html#infrastructure

Also, we have discussed merging repos into one repo. We could do this for most of the core repos, but still may have the need for some separate repos. So after merging repos there still may be a need to clearly document what individual components make up a version of OnDemand.
