### 2018-04-02 #622 @ericfranz
Thank you for the comments! I opened three issues in https://github.com/OSC/Open-OnDemand for further discussion.

1. We have an rpm we are using at OSC now and will be the default installation for OnDemand 1.3 released in April. I opened https://github.com/OSC/Open-OnDemand/issues/31 to address the problem of not advertising our development roadmap.
2. For each OnDemand release we document all the corresponding versions in the release notes (see [1.2 example](https://osc.github.io/ood-documentation/release-1.2/release-notes/v1.2-release-notes.html#infrastructure)). I opened https://github.com/OSC/Open-OnDemand/issues/32 to track the need to put this list in a more prominent location. We are also looking to combine some repos into one in the future to help with this.
3. We have some unit test coverage for components which target developers, but have no system tests for installation validation which would target admins. I opened an issue https://github.com/OSC/Open-OnDemand/issues/33 to discuss that further.
