### 2018-04-02 OSC/Open-OnDemand#31 @treydock
Minor correction, the URL for 1.3 release will be https://yum.osc.edu/ondemand/1.3, notice use of `1.3` and not `stable`.  This means that people won't accidentally be upgraded to `1.4` when that is released.
