### 2018-04-02 OSC/Open-OnDemand#33 @smgallo
@ericfranz I think that a good start would be a way to test the core infrastructure and authentication pieces. It sounds like you've already given this some thought with the job adapters. We have a growing set of tests that we use during development and have been considering shipping a small subset of these with our software to help user's verify installation. Also, one thing that we've been discussing to assist with support requests is to provide a script that a user can run to collects various bits of information from the installation, query REST endpoints, query the database, etc. and then provide that information back to us when they submit a support ticket. The output of that might be used to verify the installation as well.
