### 2018-04-02 #622 @smgallo
Thanks @ericfranz, I'll follow up on those issues to provide further input as needed. @danielskatz that addresses any comments that I had.
