### 2018-04-04 #622 @marpierc
@danielskatz what are the requirements for the Software Paper (paper.md)? The paper.md file for the project is not a tech/scientific paper in the usual sense.
