### 2018-04-04 #622 @danielskatz
@marpierc - I think I saw a comment from you about the paper and what it should contain, though I no longer see it, so perhaps it was deleted?  But in any case, see http://joss.theoj.org/about#author_guidelines (what should my paper contain)
