### 2018-04-11 #622 @marpierc
I second @smgallo comments. The major issues with the current release are the complicated installation process and the need for more automated testing. The OOD team should also consider how to handle third party contributions (bug fixes, add-ons, etc).
