### 2018-04-11 #622 @danielskatz
Thanks @marpierc

Just to make the status of this review clear, I think the ball is now with @ericfranz to address these issues and respond.
