### 2018-04-17 ce699d7
update CONTRIBUTING with proper PR workflow
