### 2018-04-17 019139c
add rpm and docs repo links
