### 2018-04-17 ae7846a
Update CONTRIBUTING with links to OOD website
