### 2018-04-17 OSC/Open-OnDemand#33 @ericfranz
Adding some system tests to verify the job configuration does seem straight forward:

   1. add a test to submit a short job, verify it was submitted, appears in the queue, and that it outputs an expected result value
  2. add a test to get a status of all the running jobs and verify that the count is > 0 and certain values that all adapters will return are populated with non-null data

As far as testing authentication piece, I'm not sure exactly how to proceed. Since Apache can be configured different authentication mechanisms variations include:

1.  Basic Auth
2. Federated authentication using Shibboleth or Open ID Connect
   1. Federated authentication we have cases where the OIDC provider is a Keycloak installation configured with ldap (so a single web form is presented to the user) and
   2. Federated authentication  with a discover page to choose an OIDC provider (like CILogon)
   3. Federated authentication like Shibboleth that require a two factor authentication step

I could imagine a test to verify Basic Auth + user mapping but I'm not sure about the other cases. And since the other cases are recommended over Basic Auth that would limit the usefulness of such a test. Since in OnDemand authentication is two parts: Apache Auth and user mapping, maybe there would be a way to test the user mapping configuration.
