### 2018-04-17 OSC/Open-OnDemand#33 @ericfranz
@smgallo @marpierc As far as JOSS submission goes, I wanted to let you know that OnDemand components do have varying degrees of test coverage - its just tests that are useful for developers, not admins installing OnDemand (examples: https://github.com/OSC/ood_core/tree/master/spec and https://github.com/OSC/ood-dashboard/tree/master/test).

What mentioned in this issue would be required or sufficient for JOSS publication?
