### 2018-04-17 OSC/Open-OnDemand#33 @smgallo
@ericfranz I think that the two items you mentioned above would be sufficient for this submission.

> Adding some system tests to verify the job configuration does seem straight forward:
> - add a test to submit a short job, verify it was submitted, appears in the queue, and that it outputs an expected result value
> - add a test to get a status of all the running jobs and verify that the count is > 0 and certain values that all adapters will return are populated with non-null data
