### 2018-04-17 6afbff6
Remove mention of "functional core" for now
