### 2018-04-22 #622 @danielskatz
👋 @ericfranz - please let us know when you have addressed these issues.
