### 2018-05-01 OSC/Open-OnDemand#33 @ericfranz
We pulled in this PR to master: https://github.com/OSC/ood-dashboard/pull/360. This will be included in the 1.3 release next week.

This adds a set of Rake tasks that can be executed via the shell on the web node. For each cluster defined in the cluster config that has a valid jobs configuration section, a rake task will be auto-generated to test submitting and checking the status of a job and a single rake task is also created that will run all of the cluster specific rake tasks - so you can execute them all, or individually. The tasks will print out the arguments sent to execve when submitting the job that should be helpful with debugging.

In the future we would like to expand this to possibly verify the installation of dependencies for interactive apps and provide a web interface to this, similar to Doku's install.php  https://www.dokuwiki.org/install.
