### 2018-05-01 #622 @ericfranz
I believe all issues have been addressed:

1. The 1.3 release next week will include deployment via rpm to address the complicated installation process
2. We added a set of auto-generated Rake tasks to test and verify job submission from the web node for each configured cluster (https://github.com/OSC/Open-OnDemand/issues/33 and https://github.com/OSC/ood-dashboard/pull/360) which will be included in the 1.3 release. This should address concerns for tests to help admins verify the installation.
3. We added a CONTRIBUTING.md file to the root of the OnDemand repo to address third party contributions.

Assuming everything goes well this week, OnDemand 1.3 will be released next Monday. Is there a need to wait for this release or can we proceed with the JOSS publication?
