### 2018-05-01 #622 @danielskatz
If the current software meets what reviewers think is needed, we can proceed.

At this point, we need @marpierc and @smgallo to examine this again and see if they can check off all the items in their checklists.
