### 2018-05-07 #622 @danielskatz
@marpierc - can you verify that the remaining point has been addressed and check off the remaining item on your list?
