### 2018-05-10 #622 @ericfranz
@arfon would it be sufficient to upload a zip or a tarball of the related rpms at https://yum.osc.edu/ondemand/1.3/ to figshare? (this is the latest release of OnDemand just released this week, which include suggested modifications). If not, what is your recommendation for how to package in a single tarball software that is spread across several repos and whose installation results in files being placed in the appropriate locations following the Filesystem Hierarchy Standard?

Other approaches I am considering include:

1. unpack the 1.3 rpms, adding a README at the root for how to proceed with the installation, and tar the parent directory
2. get tarballs of each git repo, placing them in the appropriate location in the filesystem that an rpm installation would place them, adding a README at the root of how to proceed with installation
