### 2018-05-14 #622 @ericfranz
@arfon see https://doi.org/10.6084/m9.figshare.6265514.v1

This tarball is also attached to the packaging repo https://github.com/OSC/ondemand/releases/tag/v1.3.6 and is all of the source for the core infrastructure and apps. The README.md in the tarball is markdown formatted and provides instructions for building from source, where to get help when stuck, and recommends using the RPM based installation.
