### 2018-05-15 #622 @arfon
@marpierc, @smgallo many thanks for your reviews here and to @danielskatz for editing this submission ✨

@ericfranz - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00622 :zap: 🚀:boom:
