### 2018-06-29 #622 @ericfranz
NSF would like us to add an acknowledgement to the paper. After updating the paper.md file to contain an Acknowledgements section similar to what I see in other JOSS publications, what is the next step to get the PDF to be regenerated? https://github.com/OSC/Open-OnDemand/issues/44
