### 2018-07-02 #622 @ericfranz
@arfon ok the paper.md has been updated
