### 2018-07-03 #622 @arfon
> @arfon ok the paper.md has been updated

I've updated the PDF too. This may take a few hours to show up on the live site due to the caching we do.
